+++ CHANGELOG.rst
@@ -14,6 +14,15 @@ Change Log
 Unreleased
 **********
 
+0.1.22 – 2025-10-23
+*******************
+
+Added
+=====
+
+*  feat: Optimize data migration command by implementing bulk inserts for improved performance.
+*  feat: Add management command to truncate non-empty destination tables before data migration.
+
 0.1.21 – 2025-10-22
 *******************
 
@@ -21,6 +30,7 @@ Added
 =====
 
 *  Upgrade Python Requirements
+*  fix: Convert UUIDField columns to uuid type for MariaDB
 
 0.1.20 – 2025-10-19
 *******************